Refactor handling AI requests, update User entity with more clear fields
